Pull vod-commands out into separate command Group
